+++ package.json
@@ -198,6 +198,7 @@
     "hoist-non-react-statics": "~3.3.2",
     "lodash": "~4.17.15",
     "raf": "~3.4.1",
+    "rc-virtual-list": "^3.4.2",
     "react-popper": "~2.2.5",
     "react-transition-group": "~4.4.1",
     "tdesign-icons-react": "~0.0.7",
