+++ site/test-coverage.js
@@ -1,5 +1,5 @@
 export default {
-      "unit": {"util":"34.29%","addon":"92.86%","addon/Example":"100.00%","affix":"59.26%","affix/Example":"88.24%","alert":"100.00%","alert/Example":"100.00%","anchor":"60.58%","anchor/Example":"89.47%","avatar":"96.88%","avatar/Example":"100.00%","badge":"100.00%","badge/Example":"100.00%","breadcrumb":"93.94%","breadcrumb/Example":"93.75%","button":"100.00%","button/Example":"100.00%","calendar":"89.81%","calendar/Example":"81.08%","card":"0.00%","cascader":"49.11%","cascader/Example":"77.45%","checkbox":"68.85%","checkbox/Example":"71.43%","comment":"100.00%","comment/Example":"92.31%","common":"75.56%","configProvider":"80.00%","datePicker":"22.72%","datePicker/Example":"68.00%","dialog":"51.89%","dialog/Example":"42.66%","divider":"100.00%","divider/Example":"100.00%","drawer":"66.99%","drawer/Example":"63.24%","dropdown":"23.66%","dropdown/Example":"72.73%","form":"64.29%","form/Example":"42.17%","grid":"83.33%","grid/Example":"100.00%","icon/Example":"0.00%","input":"89.04%","inputNumber":"65.91%","inputNumber/Example":"0.00%","input/Example":"52.94%","layout":"100.00%","layout/Example":"100.00%","list":"96.67%","list/Example":"77.92%","loading":"70.42%","loading/Example":"58.97%","locale":"56.52%","menu":"73.38%","menu/Example":"62.34%","message":"20.00%","message/Example":"0.00%","notification":"91.79%","notification/Example":"47.62%","pagination":"97.89%","pagination/Example":"54.76%","popconfirm":"90.38%","popconfirm/Example":"66.67%","popup":"87.83%","popup/Example":"85.71%","progress":"4.55%","progress/Example":"0.00%","radio":"89.47%","radio/Example":"66.67%","rate":"0.00%","select":"72.76%","select/Example":"75.71%","skeleton":"0.00%","slider":"6.19%","slider/Example":"0.00%","steps":"95.83%","steps/Example":"81.82%","stickyTool":"0.00%","swiper":"82.26%","swiper/Example":"88.89%","switch":"100.00%","switch/Example":"80.00%","table":"77.37%","table/Example":"62.83%","tabs":"78.68%","tabs/Example":"58.18%","tag":"100.00%","tag/Example":"87.50%","textarea":"62.86%","textarea/Example":"40.00%","timePicker":"22.38%","timePicker/Example":"78.95%","tooltip":"85.19%","tooltip/Example":"61.29%","transfer":"68.89%","transfer/Example":"91.18%","tree":"31.02%","treeSelect":"68.36%","treeSelect/Example":"57.14%","tree/Example":"40.63%","upload":"38.69%","upload/Example":"55.22%"}, 
-      "e2e": {"util":"34.29%","addon":"92.86%","addon/Example":"100.00%","affix":"59.26%","affix/Example":"88.24%","alert":"100.00%","alert/Example":"100.00%","anchor":"60.58%","anchor/Example":"89.47%","avatar":"96.88%","avatar/Example":"100.00%","badge":"100.00%","badge/Example":"100.00%","breadcrumb":"93.94%","breadcrumb/Example":"93.75%","button":"100.00%","button/Example":"100.00%","calendar":"89.81%","calendar/Example":"81.08%","card":"0.00%","cascader":"49.11%","cascader/Example":"77.45%","checkbox":"68.85%","checkbox/Example":"71.43%","comment":"100.00%","comment/Example":"92.31%","common":"75.56%","configProvider":"80.00%","datePicker":"22.72%","datePicker/Example":"68.00%","dialog":"51.89%","dialog/Example":"42.66%","divider":"100.00%","divider/Example":"100.00%","drawer":"66.99%","drawer/Example":"63.24%","dropdown":"23.66%","dropdown/Example":"72.73%","form":"64.29%","form/Example":"42.17%","grid":"83.33%","grid/Example":"100.00%","icon/Example":"0.00%","input":"89.04%","inputNumber":"65.91%","inputNumber/Example":"0.00%","input/Example":"52.94%","layout":"100.00%","layout/Example":"100.00%","list":"96.67%","list/Example":"77.92%","loading":"70.42%","loading/Example":"58.97%","locale":"56.52%","menu":"73.38%","menu/Example":"62.34%","message":"20.00%","message/Example":"0.00%","notification":"91.79%","notification/Example":"47.62%","pagination":"97.89%","pagination/Example":"54.76%","popconfirm":"90.38%","popconfirm/Example":"66.67%","popup":"87.83%","popup/Example":"85.71%","progress":"4.55%","progress/Example":"0.00%","radio":"89.47%","radio/Example":"66.67%","rate":"0.00%","select":"72.76%","select/Example":"75.71%","skeleton":"0.00%","slider":"6.19%","slider/Example":"0.00%","steps":"95.83%","steps/Example":"81.82%","stickyTool":"0.00%","swiper":"82.26%","swiper/Example":"88.89%","switch":"100.00%","switch/Example":"80.00%","table":"77.37%","table/Example":"62.83%","tabs":"78.68%","tabs/Example":"58.18%","tag":"100.00%","tag/Example":"87.50%","textarea":"62.86%","textarea/Example":"40.00%","timePicker":"22.38%","timePicker/Example":"78.95%","tooltip":"85.19%","tooltip/Example":"61.29%","transfer":"68.89%","transfer/Example":"91.18%","tree":"31.02%","treeSelect":"68.36%","treeSelect/Example":"57.14%","tree/Example":"40.63%","upload":"38.69%","upload/Example":"55.22%"}, 
-      "ssr": {"util":"31.14%","addon":"92.86%","addon/Example":"100.00%","affix":"24.07%","affix/Example":"64.71%","alert":"82.76%","alert/Example":"100.00%","anchor":"32.85%","anchor/Example":"89.47%","avatar":"81.25%","avatar/Example":"100.00%","badge":"100.00%","badge/Example":"100.00%","breadcrumb":"93.94%","breadcrumb/Example":"93.75%","button":"94.74%","button/Example":"100.00%","calendar":"79.13%","calendar/Example":"74.77%","card":"0.00%","cascader":"35.50%","cascader/Example":"77.45%","checkbox":"63.93%","checkbox/Example":"71.43%","comment":"100.00%","comment/Example":"92.31%","common":"51.11%","configProvider":"80.00%","datePicker":"13.82%","datePicker/Example":"68.00%","dialog":"47.03%","dialog/Example":"42.66%","divider":"91.67%","divider/Example":"100.00%","drawer":"50.49%","drawer/Example":"63.24%","dropdown":"23.66%","dropdown/Example":"72.73%","form":"32.54%","form/Example":"40.96%","grid":"80.00%","grid/Example":"100.00%","icon/Example":"100.00%","input":"53.42%","inputNumber":"56.06%","inputNumber/Example":"54.55%","input/Example":"52.94%","layout":"93.02%","layout/Example":"100.00%","list":"76.67%","list/Example":"71.43%","loading":"60.56%","loading/Example":"48.72%","locale":"56.52%","menu":"73.38%","menu/Example":"62.34%","message":"43.00%","message/Example":"44.12%","notification":"35.07%","notification/Example":"47.62%","pagination":"44.21%","pagination/Example":"54.76%","popconfirm":"59.62%","popconfirm/Example":"66.67%","popup":"54.78%","popup/Example":"78.57%","progress":"95.45%","progress/Example":"100.00%","radio":"71.05%","radio/Example":"66.67%","rate":"0.00%","select":"22.07%","select/Example":"75.71%","skeleton":"0.00%","slider":"52.21%","slider/Example":"100.00%","steps":"91.67%","steps/Example":"81.82%","stickyTool":"0.00%","swiper":"62.90%","swiper/Example":"88.89%","switch":"74.07%","switch/Example":"80.00%","table":"68.81%","table/Example":"61.06%","tabs":"56.62%","tabs/Example":"58.18%","tag":"90.48%","tag/Example":"70.83%","textarea":"62.86%","textarea/Example":"40.00%","timePicker":"19.93%","timePicker/Example":"78.95%","tooltip":"62.96%","tooltip/Example":"48.39%","transfer":"62.96%","transfer/Example":"91.18%","tree":"30.61%","treeSelect":"58.19%","treeSelect/Example":"57.14%","tree/Example":"39.90%","upload":"38.44%","upload/Example":"55.22%"}
+      "unit": {"util":"34.29%","addon":"92.86%","addon/Example":"100.00%","affix":"59.26%","affix/Example":"88.24%","alert":"100.00%","alert/Example":"100.00%","anchor":"60.58%","anchor/Example":"89.47%","avatar":"96.88%","avatar/Example":"100.00%","badge":"100.00%","badge/Example":"100.00%","breadcrumb":"93.94%","breadcrumb/Example":"93.75%","button":"100.00%","button/Example":"100.00%","calendar":"89.81%","calendar/Example":"81.08%","card":"0.00%","cascader":"49.11%","cascader/Example":"77.45%","checkbox":"68.85%","checkbox/Example":"71.43%","comment":"100.00%","comment/Example":"92.31%","common":"75.56%","configProvider":"80.00%","datePicker":"22.72%","datePicker/Example":"68.00%","dialog":"51.89%","dialog/Example":"42.66%","divider":"100.00%","divider/Example":"100.00%","drawer":"66.99%","drawer/Example":"63.24%","dropdown":"23.66%","dropdown/Example":"72.73%","form":"64.29%","form/Example":"42.17%","grid":"83.33%","grid/Example":"100.00%","icon/Example":"0.00%","input":"89.04%","inputNumber":"65.91%","inputNumber/Example":"0.00%","input/Example":"52.94%","layout":"100.00%","layout/Example":"100.00%","list":"96.67%","list/Example":"79.76%","loading":"70.42%","loading/Example":"58.97%","locale":"56.52%","menu":"73.38%","menu/Example":"62.34%","message":"20.00%","message/Example":"0.00%","notification":"91.79%","notification/Example":"47.62%","pagination":"97.89%","pagination/Example":"54.76%","popconfirm":"90.38%","popconfirm/Example":"66.67%","popup":"87.83%","popup/Example":"85.71%","progress":"4.55%","progress/Example":"0.00%","radio":"89.47%","radio/Example":"66.67%","rate":"0.00%","select":"72.76%","select/Example":"75.71%","skeleton":"0.00%","slider":"6.19%","slider/Example":"0.00%","steps":"95.83%","steps/Example":"81.82%","stickyTool":"0.00%","swiper":"82.26%","swiper/Example":"88.89%","switch":"100.00%","switch/Example":"80.00%","table":"77.37%","table/Example":"62.83%","tabs":"78.68%","tabs/Example":"58.18%","tag":"100.00%","tag/Example":"87.50%","textarea":"62.86%","textarea/Example":"40.00%","timePicker":"22.38%","timePicker/Example":"78.95%","tooltip":"85.19%","tooltip/Example":"61.29%","transfer":"68.89%","transfer/Example":"91.18%","tree":"31.02%","treeSelect":"68.36%","treeSelect/Example":"57.14%","tree/Example":"40.63%","upload":"38.69%","upload/Example":"55.22%"}, 
+      "e2e": {"util":"34.29%","addon":"92.86%","addon/Example":"100.00%","affix":"59.26%","affix/Example":"88.24%","alert":"100.00%","alert/Example":"100.00%","anchor":"60.58%","anchor/Example":"89.47%","avatar":"96.88%","avatar/Example":"100.00%","badge":"100.00%","badge/Example":"100.00%","breadcrumb":"93.94%","breadcrumb/Example":"93.75%","button":"100.00%","button/Example":"100.00%","calendar":"89.81%","calendar/Example":"81.08%","card":"0.00%","cascader":"49.11%","cascader/Example":"77.45%","checkbox":"68.85%","checkbox/Example":"71.43%","comment":"100.00%","comment/Example":"92.31%","common":"75.56%","configProvider":"80.00%","datePicker":"22.72%","datePicker/Example":"68.00%","dialog":"51.89%","dialog/Example":"42.66%","divider":"100.00%","divider/Example":"100.00%","drawer":"66.99%","drawer/Example":"63.24%","dropdown":"23.66%","dropdown/Example":"72.73%","form":"64.29%","form/Example":"42.17%","grid":"83.33%","grid/Example":"100.00%","icon/Example":"0.00%","input":"89.04%","inputNumber":"65.91%","inputNumber/Example":"0.00%","input/Example":"52.94%","layout":"100.00%","layout/Example":"100.00%","list":"96.67%","list/Example":"79.76%","loading":"70.42%","loading/Example":"58.97%","locale":"56.52%","menu":"73.38%","menu/Example":"62.34%","message":"20.00%","message/Example":"0.00%","notification":"91.79%","notification/Example":"47.62%","pagination":"97.89%","pagination/Example":"54.76%","popconfirm":"90.38%","popconfirm/Example":"66.67%","popup":"87.83%","popup/Example":"85.71%","progress":"4.55%","progress/Example":"0.00%","radio":"89.47%","radio/Example":"66.67%","rate":"0.00%","select":"72.76%","select/Example":"75.71%","skeleton":"0.00%","slider":"6.19%","slider/Example":"0.00%","steps":"95.83%","steps/Example":"81.82%","stickyTool":"0.00%","swiper":"82.26%","swiper/Example":"88.89%","switch":"100.00%","switch/Example":"80.00%","table":"77.37%","table/Example":"62.83%","tabs":"78.68%","tabs/Example":"58.18%","tag":"100.00%","tag/Example":"87.50%","textarea":"62.86%","textarea/Example":"40.00%","timePicker":"22.38%","timePicker/Example":"78.95%","tooltip":"85.19%","tooltip/Example":"61.29%","transfer":"68.89%","transfer/Example":"91.18%","tree":"31.02%","treeSelect":"68.36%","treeSelect/Example":"57.14%","tree/Example":"40.63%","upload":"38.69%","upload/Example":"55.22%"}, 
+      "ssr": {"util":"31.14%","addon":"92.86%","addon/Example":"100.00%","affix":"24.07%","affix/Example":"64.71%","alert":"82.76%","alert/Example":"100.00%","anchor":"32.85%","anchor/Example":"89.47%","avatar":"81.25%","avatar/Example":"100.00%","badge":"100.00%","badge/Example":"100.00%","breadcrumb":"93.94%","breadcrumb/Example":"93.75%","button":"94.74%","button/Example":"100.00%","calendar":"79.13%","calendar/Example":"74.77%","card":"0.00%","cascader":"35.50%","cascader/Example":"77.45%","checkbox":"63.93%","checkbox/Example":"71.43%","comment":"100.00%","comment/Example":"92.31%","common":"51.11%","configProvider":"80.00%","datePicker":"13.82%","datePicker/Example":"68.00%","dialog":"47.03%","dialog/Example":"42.66%","divider":"91.67%","divider/Example":"100.00%","drawer":"50.49%","drawer/Example":"63.24%","dropdown":"23.66%","dropdown/Example":"72.73%","form":"32.54%","form/Example":"40.96%","grid":"80.00%","grid/Example":"100.00%","icon/Example":"100.00%","input":"53.42%","inputNumber":"56.06%","inputNumber/Example":"54.55%","input/Example":"52.94%","layout":"93.02%","layout/Example":"100.00%","list":"76.67%","list/Example":"73.81%","loading":"60.56%","loading/Example":"48.72%","locale":"56.52%","menu":"73.38%","menu/Example":"62.34%","message":"43.00%","message/Example":"44.12%","notification":"35.07%","notification/Example":"47.62%","pagination":"44.21%","pagination/Example":"54.76%","popconfirm":"59.62%","popconfirm/Example":"66.67%","popup":"54.78%","popup/Example":"78.57%","progress":"95.45%","progress/Example":"100.00%","radio":"71.05%","radio/Example":"66.67%","rate":"0.00%","select":"22.07%","select/Example":"75.71%","skeleton":"0.00%","slider":"52.21%","slider/Example":"100.00%","steps":"91.67%","steps/Example":"81.82%","stickyTool":"0.00%","swiper":"62.90%","swiper/Example":"88.89%","switch":"74.07%","switch/Example":"80.00%","table":"68.81%","table/Example":"61.06%","tabs":"56.62%","tabs/Example":"58.18%","tag":"90.48%","tag/Example":"70.83%","textarea":"62.86%","textarea/Example":"40.00%","timePicker":"19.93%","timePicker/Example":"78.95%","tooltip":"62.96%","tooltip/Example":"48.39%","transfer":"62.96%","transfer/Example":"91.18%","tree":"30.61%","treeSelect":"58.19%","treeSelect/Example":"57.14%","tree/Example":"39.90%","upload":"38.44%","upload/Example":"55.22%"}
     }
